feat: restructure Arr and AppStorage
