+++ faderpunk/src/app.rs
@@ -1,14 +1,9 @@
 use embassy_rp::clocks::RoscRng;
-use embassy_sync::{
-    blocking_mutex::raw::NoopRawMutex, channel::Sender, mutex::Mutex, pubsub::Subscriber,
-    signal::Signal, watch::Receiver,
-};
+use embassy_sync::{blocking_mutex::raw::NoopRawMutex, mutex::Mutex, signal::Signal};
 use embassy_time::{Duration, Timer};
-use heapless::Vec;
 use max11300::config::{ConfigMode3, ConfigMode5, ConfigMode7, ADCRANGE, AVR, DACRANGE, NSAMPLES};
 use midly::{live::LiveEvent, num::u4, MidiMessage};
 use portable_atomic::Ordering;
-use postcard::{from_bytes, to_slice};
 use rand::Rng;
 
 use config::Curve;
@@ -16,18 +11,11 @@ use libfp::{
     constants::{CHAN_LED_MAP, CURVE_EXP, CURVE_LOG},
     utils::scale_bits_12_7,
 };
-use serde::{
-    de::{DeserializeOwned, Error as DeError},
-    ser::Error as SerError,
-    Deserialize, Deserializer, Serialize, Serializer,
-};
 
 use crate::{
-    storage::AppStorageAddress,
     tasks::{
         buttons::BUTTON_PRESSED,
         clock::{ClockSubscriber, CLOCK_PUBSUB},
-        fram::{request_data, write_data, ReadOperation, WriteOperation, FRAM_WRITE_BUF},
         leds::LED_VALUES,
         max::{MaxCmd, MaxConfig, MaxSender, MAX_VALUES_ADC, MAX_VALUES_DAC, MAX_VALUES_FADER},
         midi::MidiSender,
@@ -35,7 +23,10 @@ use crate::{
     EventPubSubChannel, InputEvent,
 };
 
-pub use crate::tasks::clock::ClockEvent;
+pub use crate::{
+    storage::{AppStorage, Arr, ManagedStorage, ParamSlot, ParamStore},
+    tasks::clock::ClockEvent,
+};
 
 // TODO: This will be refactored using an allocator
 const BYTES_PER_VALUE_SET: u32 = 1000;
@@ -54,56 +45,6 @@ pub enum Led {
     Top,
     Bottom,
     Button,
-}
-
-// TODO: Move to storage
-#[derive(Clone, Copy)]
-pub struct Arr<T: Sized + Copy + Default, const N: usize>(pub [T; N]);
-
-impl<T: Sized + Copy + Default, const N: usize> Default for Arr<T, N> {
-    fn default() -> Self {
-        Self([T::default(); N])
-    }
-}
-
-impl<T, const N: usize> Serialize for Arr<T, N>
-where
-    T: Serialize + Sized + Copy + Default,
-{
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let vec = Vec::<T, N>::from_slice(&self.0).unwrap();
-        vec.serialize(serializer)
-    }
-}
-
-impl<'de, T, const N: usize> Deserialize<'de> for Arr<T, N>
-where
-    T: Deserialize<'de> + Sized + Copy + Default,
-{
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let vec = Vec::<T, N>::deserialize(deserializer)?;
-        if vec.len() != N {
-            return Err(D::Error::invalid_length(
-                vec.len(),
-                &"an array of exact length N",
-            ));
-        }
-        let mut arr = [T::default(); N];
-        arr.copy_from_slice(vec.as_slice()); // Safe due to length check above
-        Ok(Arr(arr))
-    }
-}
-
-impl<T: Sized + Copy + PartialEq + Default, const N: usize> PartialEq for Arr<T, N> {
-    fn eq(&self, other: &Self) -> bool {
-        self.0 == other.0
-    }
 }
 
 #[derive(Clone, Copy)]
@@ -456,48 +397,6 @@ impl<T: Sized + Copy + Default> Default for Global<T> {
     }
 }
 
-impl<T> Serialize for Global<T>
-where
-    T: Sized + Copy + Serialize, // T must also be Serialize
-{
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        match self.inner.try_lock() {
-            Ok(guard) => {
-                // Dereference the guard to get &T and serialize it.
-                // The Copy trait on T means *guard would yield T by value,
-                // but &*guard or simply guard (if it derefs to &T) is what serialize expects.
-                // (*guard) also works as serialize takes &self.
-                (*guard).serialize(serializer)
-            }
-            Err(_e) => {
-                // _e is typically embassy_sync::mutex::TryLockError
-                // If the lock can't be acquired synchronously, serialization fails.
-                Err(S::Error::custom("could not lock mutex for serialization"))
-            }
-        }
-    }
-}
-
-impl<'de, T> Deserialize<'de> for Global<T>
-where
-    T: Sized + Copy + Deserialize<'de>, // T must also be Deserialize
-{
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        // Deserialize T directly.
-        let value = T::deserialize(deserializer)?;
-        // Create a new Mutex holding the deserialized T.
-        Ok(Global {
-            inner: Mutex::new(value),
-        })
-    }
-}
-
 pub struct Die {
     rng: RoscRng,
 }
@@ -657,82 +556,5 @@ impl<const N: usize> App<N> {
     pub async fn exit_handler(&self, exit_signal: &'static Signal<NoopRawMutex, bool>) {
         exit_signal.wait().await;
         self.reset().await;
-    }
-}
-
-pub trait AppStorage:
-    Serialize + for<'de> Deserialize<'de> + Default + Send + Sync + 'static
-{
-}
-
-pub struct ManagedStorage<S: AppStorage> {
-    app_id: u8,
-    inner: Mutex<NoopRawMutex, S>,
-    start_channel: usize,
-}
-
-impl<S: AppStorage> ManagedStorage<S> {
-    pub fn new(app_id: u8, start_channel: usize) -> Self {
-        Self {
-            app_id,
-            inner: Mutex::new(S::default()),
-            start_channel,
-        }
-    }
-
-    pub async fn load(&self, scene: Option<u8>) {
-        let address = AppStorageAddress::new(self.start_channel, scene);
-        let op = ReadOperation::new(address.into());
-        if let Ok(data) = request_data(op).await {
-            if data.is_empty() {
-                return;
-            }
-            if let Ok(val) = from_bytes::<S>(&data[1..]) {
-                if data[0] != self.app_id {
-                    return;
-                }
-                let mut inner = self.inner.lock().await;
-                *inner = val;
-            }
-        }
-    }
-
-    pub async fn save(&self, scene: Option<u8>) {
-        let mut data = FRAM_WRITE_BUF.lock().await;
-
-        data[0] = self.app_id;
-        let inner = self.inner.lock().await;
-        let len = to_slice(&*inner, &mut data[1..]).unwrap().len();
-
-        let address = AppStorageAddress::new(self.start_channel, scene);
-        if let Ok(op) = WriteOperation::try_new(address.into(), len + 1) {
-            write_data(op).await.unwrap();
-        }
-    }
-
-    pub async fn query<F, R>(&self, accessor: F) -> R
-    where
-        F: FnOnce(&S) -> R,
-    {
-        let guard = self.inner.lock().await;
-        accessor(&*guard)
-    }
-
-    pub async fn modify<F, R>(&self, modifier: F) -> R
-    where
-        F: FnOnce(&mut S) -> R,
-    {
-        let mut guard = self.inner.lock().await;
-        modifier(&mut *guard)
-    }
-
-    pub async fn modify_and_save<F, R>(&self, modifier: F, scene: Option<u8>) -> R
-    where
-        F: FnOnce(&mut S) -> R,
-    {
-        let mut guard = self.inner.lock().await;
-        let s = modifier(&mut *guard);
-        self.save(scene).await;
-        s
     }
 }
