+++ faderpunk/src/storage.rs
@@ -4,11 +4,14 @@ use config::{FromValue, GlobalConfig, Value, APP_MAX_PARAMS};
 use embassy_sync::{blocking_mutex::raw::NoopRawMutex, mutex::Mutex};
 use heapless::Vec;
 use postcard::{from_bytes, to_slice};
-use serde::{de::DeserializeOwned, Serialize};
+use serde::{
+    de::{DeserializeOwned, Error as DeError},
+    Deserialize, Deserializer, Serialize, Serializer,
+};
 
 use crate::tasks::{
     configure::{AppParamCmd, APP_PARAM_CHANNEL, APP_PARAM_SIGNALS},
-    fram::{request_data, write_data, ReadOperation, WriteOperation, FRAM_WRITE_BUF, MAX_DATA_LEN},
+    fram::{request_data, write_data, ReadOperation, WriteOperation, FRAM_WRITE_BUF},
 };
 
 const GLOBAL_CONFIG_RANGE: Range<u32> = 0..1_024;
@@ -38,6 +41,81 @@ pub async fn load_global_config() -> GlobalConfig {
         .ok()
         .and_then(|data| from_bytes::<GlobalConfig>(&data).ok())
         .unwrap_or_else(GlobalConfig::new)
+}
+
+#[derive(Clone, Copy)]
+pub struct Arr<T: Sized + Copy + Default, const N: usize>([T; N]);
+
+impl<T: Sized + Copy + Default, const N: usize> Default for Arr<T, N> {
+    fn default() -> Self {
+        Self([T::default(); N])
+    }
+}
+
+impl<T: Sized + Copy + Default, const N: usize> Arr<T, N> {
+    pub fn new(initial: [T; N]) -> Self {
+        Self(initial)
+    }
+
+    #[inline(always)]
+    pub fn at(&self, idx: usize) -> T {
+        self.0[idx]
+    }
+
+    #[inline(always)]
+    pub fn set_at(&mut self, idx: usize, value: T) {
+        self.0[idx] = value;
+    }
+
+    #[inline(always)]
+    pub fn get(&self) -> [T; N] {
+        self.0
+    }
+
+    #[inline(always)]
+    pub fn set(&mut self, value: [T; N]) {
+        self.0 = value;
+    }
+}
+
+impl<T, const N: usize> Serialize for Arr<T, N>
+where
+    T: Serialize + Sized + Copy + Default,
+{
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let vec = Vec::<T, N>::from_slice(&self.0).unwrap();
+        vec.serialize(serializer)
+    }
+}
+
+impl<'de, T, const N: usize> Deserialize<'de> for Arr<T, N>
+where
+    T: Deserialize<'de> + Sized + Copy + Default,
+{
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let vec = Vec::<T, N>::deserialize(deserializer)?;
+        if vec.len() != N {
+            return Err(D::Error::invalid_length(
+                vec.len(),
+                &"an array of exact length N",
+            ));
+        }
+        let mut arr = [T::default(); N];
+        arr.copy_from_slice(vec.as_slice()); // Safe due to length check above
+        Ok(Arr(arr))
+    }
+}
+
+impl<T: Sized + Copy + PartialEq + Default, const N: usize> PartialEq for Arr<T, N> {
+    fn eq(&self, other: &Self) -> bool {
+        self.0 == other.0
+    }
 }
 
 #[derive(Clone, Copy)]
@@ -249,5 +327,82 @@ where
     pub async fn get(&self) -> T {
         let value = self.values.get(self.index).await;
         T::from_value(value)
+    }
+}
+
+pub trait AppStorage:
+    Serialize + for<'de> Deserialize<'de> + Default + Send + Sync + 'static
+{
+}
+
+pub struct ManagedStorage<S: AppStorage> {
+    app_id: u8,
+    inner: Mutex<NoopRawMutex, S>,
+    start_channel: usize,
+}
+
+impl<S: AppStorage> ManagedStorage<S> {
+    pub fn new(app_id: u8, start_channel: usize) -> Self {
+        Self {
+            app_id,
+            inner: Mutex::new(S::default()),
+            start_channel,
+        }
+    }
+
+    pub async fn load(&self, scene: Option<u8>) {
+        let address = AppStorageAddress::new(self.start_channel, scene);
+        let op = ReadOperation::new(address.into());
+        if let Ok(data) = request_data(op).await {
+            if data.is_empty() {
+                return;
+            }
+            if let Ok(val) = from_bytes::<S>(&data[1..]) {
+                if data[0] != self.app_id {
+                    return;
+                }
+                let mut inner = self.inner.lock().await;
+                *inner = val;
+            }
+        }
+    }
+
+    pub async fn save(&self, scene: Option<u8>) {
+        let mut data = FRAM_WRITE_BUF.lock().await;
+
+        data[0] = self.app_id;
+        let inner = self.inner.lock().await;
+        let len = to_slice(&*inner, &mut data[1..]).unwrap().len();
+
+        let address = AppStorageAddress::new(self.start_channel, scene);
+        if let Ok(op) = WriteOperation::try_new(address.into(), len + 1) {
+            write_data(op).await.unwrap();
+        }
+    }
+
+    pub async fn query<F, R>(&self, accessor: F) -> R
+    where
+        F: FnOnce(&S) -> R,
+    {
+        let guard = self.inner.lock().await;
+        accessor(&*guard)
+    }
+
+    pub async fn modify<F, R>(&self, modifier: F) -> R
+    where
+        F: FnOnce(&mut S) -> R,
+    {
+        let mut guard = self.inner.lock().await;
+        modifier(&mut *guard)
+    }
+
+    pub async fn modify_and_save<F, R>(&self, modifier: F, scene: Option<u8>) -> R
+    where
+        F: FnOnce(&mut S) -> R,
+    {
+        let mut guard = self.inner.lock().await;
+        let s = modifier(&mut *guard);
+        self.save(scene).await;
+        s
     }
 }
